Add choco_updater script

+++ windows/bin/choco.ps1
@@ -5,10 +5,18 @@ Param(
 $chocoInstall = 'bin\choco\tools\chocolateyInstall.ps1'
 $path = "build\installers\headset-$version-setup.exe"
 
+if (!(Test-Path -Path "$path")) {
+  exit 1
+}
+
 $sha256 = (CertUtil -hashfile "$path" SHA256)[1].Replace(' ', '')
 
 $choco = Get-Content $chocoInstall
 $choco[8] = "  checksum       = '$sha256'"
 $choco | Out-File -filepath "$chocoInstall" -Encoding UTF8
+
+if (!(Test-Path -Path "$PWD\build")) {
+  New-Item -ItemType Directory -Force -Path "$PWD\build"
+}
 
 choco pack 'bin\choco\headset.nuspec' -outputdirectory 'build'
